requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
flask-socketio==5.3.4
python-socketio==5.9.0
werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.5
PyJWT==2.8.0
gunicorn==21.2.0
eventlet==0.33.3
redis==5.0.0